Delete initialize_temperature
