+++ src/dynamics/dynamics.hpp
@@ -14,8 +14,8 @@
 #include "../simulation/spacecraft/structure/structure.hpp"
 #include "dynamics/attitude/initialize_attitude.hpp"
 #include "dynamics/orbit/initialize_orbit.hpp"
-#include "dynamics/thermal/initialize_temperature.hpp"
 #include "dynamics/thermal/node.hpp"
+#include "dynamics/thermal/temperature.hpp"
 
 class RelativeInformation;
 class LocalEnvironment;
