+++ src/dynamics/thermal/temperature.hpp
@@ -150,4 +150,13 @@ class Temperature : public ILoggable {
   void PrintParams(void);
 };
 
+/**
+ * @fn InitTemperature
+ * @brief Initialize Temperature object from csv file
+ * @param[in] file_name: Directory of thermal input files
+ * @param[in] rk_prop_step_s: time step interval for temperature propagation integration
+ * @return Temperature*
+ */
+Temperature* InitTemperature(const std::string file_name, const double rk_prop_step_s, const SolarRadiationPressureEnvironment* srp_environment);
+
 #endif  // S2E_DYNAMICS_THERMAL_TEMPERATURE_HPP_
